refactor: unify auth context and update tests
