--- __tests__/auth-form.test.tsx
@@ -1,60 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import { AuthForm } from '@/components/auth-form';
-import { vi } from 'vitest';
-import { AuthProvider } from '@/lib/firebase/auth-context';
-
-// Mock Next.js router
-vi.mock('next/navigation', () => ({
-  useRouter: () => ({
-    push: vi.fn(),
-    replace: vi.fn(),
-    refresh: vi.fn(),
-  }),
-  usePathname: () => '/',
-}));
-
-// Mock Firebase
-vi.mock('firebase/app', () => ({
-  initializeApp: vi.fn(),
-  getApps: vi.fn(() => []),
-  getApp: vi.fn(),
-}));
-
-vi.mock('firebase/auth', () => ({
-  getAuth: vi.fn(() => ({
-    currentUser: null,
-  })),
-  signInWithEmailAndPassword: vi.fn(),
-  createUserWithEmailAndPassword: vi.fn(),
-  signOut: vi.fn(),
-  onAuthStateChanged: vi.fn((auth, callback) => {
-    callback(null);
-    return () => {};
-  }),
-}));
-
-vi.mock('firebase/firestore', () => ({
-  getFirestore: vi.fn(),
-  doc: vi.fn(),
-  getDoc: vi.fn(),
-  setDoc: vi.fn(),
-  collection: vi.fn(),
-  query: vi.fn(),
-  where: vi.fn(),
-  getDocs: vi.fn(),
-}));
-
-describe('AuthForm', () => {
-  it('renders email and password fields', () => {
-    const mockSetMfaResolver = vi.fn();
-    
-    render(
-      <AuthProvider>
-        <AuthForm setMfaResolver={mockSetMfaResolver} />
-      </AuthProvider>
-    );
-    
-    expect(screen.getByLabelText('Email Address')).toBeInTheDocument();
-    expect(screen.getByLabelText('Password')).toBeInTheDocument();
-  });
-});
