--- lib/firebase/auth-context.tsx
@@ -1,95 +1,0 @@
-'use client';
-
-import { 
-  createContext, 
-  useContext, 
-  useEffect, 
-  useState, 
-  type ReactNode 
-} from 'react';
-import { useRouter } from 'next/navigation';
-import { 
-  type User, 
-  onIdTokenChanged, 
-  signOut as firebaseSignOut,
-  createUserWithEmailAndPassword,
-  signInWithEmailAndPassword,
-  signInWithPopup,
-  GoogleAuthProvider,
-  sendPasswordResetEmail
-} from 'firebase/auth';
-import { auth } from '@/lib/firebase/client';
-
-interface AuthContextType {
-  user: User | null;
-  loading: boolean;
-  signInWithEmail: typeof signInWithEmailAndPassword;
-  signInWithGoogle: () => Promise<void>;
-  createUser: typeof createUserWithEmailAndPassword;
-  sendPasswordReset: typeof sendPasswordResetEmail;
-  signOut: typeof firebaseSignOut;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export function AuthProvider({ children }: { children: ReactNode }) {
-  const [user, setUser] = useState<User | null>(null);
-  const [loading, setLoading] = useState(true);
-  const router = useRouter();
-
-  useEffect(() => {
-    const unsubscribe = onIdTokenChanged(auth, async (user) => {
-      setUser(user);
-      setLoading(false);
-      if (user) {
-        const idToken = await user.getIdToken();
-        // Send the token to your server to create a session cookie
-        await fetch('/api/auth/session', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({ idToken }),
-        });
-        router.push('/'); // Redirect to home on successful login
-      }
-    });
-
-    return () => unsubscribe();
-  }, [router]);
-
-  const signInWithGoogle = async () => {
-    const provider = new GoogleAuthProvider();
-    await signInWithPopup(auth, provider);
-  };
-  
-  const signOut = async () => {
-    await firebaseSignOut(auth);
-    await fetch('/api/auth/session', { method: 'DELETE' });
-    router.push('/login');
-  };
-
-  const value = {
-    user,
-    loading,
-    signInWithEmail: signInWithEmailAndPassword.bind(null, auth),
-    signInWithGoogle,
-    createUser: createUserWithEmailAndPassword.bind(null, auth),
-    sendPasswordReset: sendPasswordResetEmail.bind(null, auth),
-    signOut
-  };
-
-  return (
-    <AuthContext.Provider value={value}>
-      {!loading && children}
-    </AuthContext.Provider>
-  );
-}
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
-};
